refactor : order-test 도메인 분리
