+++ blackWhiteDeliveryDriver/src/test/java/com/sparta/blackwhitedeliverydriver/order/dto/OrderUpdateRequestDtoTest.java
@@ -1,8 +1,8 @@
-package com.sparta.blackwhitedeliverydriver.dto;
+package com.sparta.blackwhitedeliverydriver.order.dto;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
-import com.sparta.blackwhitedeliverydriver.entity.OrderStatusEnum;
+import com.sparta.blackwhitedeliverydriver.order.entity.OrderStatusEnum;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
