+++ blackWhiteDeliveryDriver/src/test/java/com/sparta/blackwhitedeliverydriver/order/service/OrderServiceTest.java
@@ -1,4 +1,4 @@
-package com.sparta.blackwhitedeliverydriver.service;
+package com.sparta.blackwhitedeliverydriver.order.service;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -8,27 +8,27 @@ import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.sparta.blackwhitedeliverydriver.dto.OrderAddRequestDto;
-import com.sparta.blackwhitedeliverydriver.dto.OrderGetDetailResponseDto;
-import com.sparta.blackwhitedeliverydriver.dto.OrderGetResponseDto;
-import com.sparta.blackwhitedeliverydriver.dto.OrderResponseDto;
-import com.sparta.blackwhitedeliverydriver.dto.OrderUpdateRequestDto;
-import com.sparta.blackwhitedeliverydriver.entity.Basket;
-import com.sparta.blackwhitedeliverydriver.entity.Order;
-import com.sparta.blackwhitedeliverydriver.entity.OrderProduct;
-import com.sparta.blackwhitedeliverydriver.entity.OrderStatusEnum;
-import com.sparta.blackwhitedeliverydriver.entity.OrderTypeEnum;
-import com.sparta.blackwhitedeliverydriver.entity.Product;
-import com.sparta.blackwhitedeliverydriver.entity.Store;
-import com.sparta.blackwhitedeliverydriver.entity.User;
-import com.sparta.blackwhitedeliverydriver.entity.UserRoleEnum;
-import com.sparta.blackwhitedeliverydriver.exception.ExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.exception.OrderExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.repository.BasketRepository;
-import com.sparta.blackwhitedeliverydriver.repository.OrderProductRepository;
-import com.sparta.blackwhitedeliverydriver.repository.OrderRepository;
-import com.sparta.blackwhitedeliverydriver.repository.StoreRepository;
-import com.sparta.blackwhitedeliverydriver.repository.UserRepository;
+import com.sparta.blackwhitedeliverydriver.order.dto.OrderAddRequestDto;
+import com.sparta.blackwhitedeliverydriver.order.dto.OrderGetDetailResponseDto;
+import com.sparta.blackwhitedeliverydriver.order.dto.OrderResponseDto;
+import com.sparta.blackwhitedeliverydriver.order.dto.OrderUpdateRequestDto;
+import com.sparta.blackwhitedeliverydriver.bascket.entity.Basket;
+import com.sparta.blackwhitedeliverydriver.order.entity.Order;
+import com.sparta.blackwhitedeliverydriver.order.entity.OrderProduct;
+import com.sparta.blackwhitedeliverydriver.order.entity.OrderStatusEnum;
+import com.sparta.blackwhitedeliverydriver.order.entity.OrderTypeEnum;
+import com.sparta.blackwhitedeliverydriver.product.entity.Product;
+import com.sparta.blackwhitedeliverydriver.pay.service.PayService;
+import com.sparta.blackwhitedeliverydriver.store.entity.Store;
+import com.sparta.blackwhitedeliverydriver.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.user.entity.UserRoleEnum;
+import com.sparta.blackwhitedeliverydriver.user.exception.ExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.order.exception.OrderExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.bascket.repository.BasketRepository;
+import com.sparta.blackwhitedeliverydriver.order.repository.OrderProductRepository;
+import com.sparta.blackwhitedeliverydriver.order.repository.OrderRepository;
+import com.sparta.blackwhitedeliverydriver.store.repository.StoreRepository;
+import com.sparta.blackwhitedeliverydriver.user.repository.UserRepository;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
